+++ components/general/Appbar.tsx
@@ -1,12 +1,12 @@
 import Link from 'next/link';
 import React, { useContext } from 'react';
-import Button from './atom/Button';
-import Container from './atom/Container';
-import Navbar from './atom/Navbar';
+import Button from '../atom/Button';
+import Container from '../atom/Container';
+import Navbar from '../atom/Navbar';
 import { AuthContext } from './CognitoAuthProvider';
 import Logo from './Logo';
-import NavSearch from './NavSearch';
-import SearchButton from './SearchButton';
+import NavSearch from '../search/NavSearch';
+import SearchButton from '../search/SearchButton';
 
 const Appbar: React.FC = () => {
   const { hasResolved, user, signOut } = useContext(AuthContext);
