+++ components/search/NavSearch.tsx
@@ -2,10 +2,10 @@ import classNames from 'classnames';
 import React, { HTMLAttributes, useEffect, useRef, useState } from 'react';
 import { FiSearch } from 'react-icons/fi';
 import { MixpanelConsumer } from 'react-mixpanel';
-import fetchSearchResults from '../functions/fetchSearchResults';
-import Expand from './atom/Expand';
-import Ripple from './atom/Ripple';
-import Loader from './Loader';
+import fetchSearchResults from '../../functions/fetchSearchResults';
+import Expand from '../atom/Expand';
+import Ripple from '../atom/Ripple';
+import Loader from '../general/Loader';
 import SearchResult, { CourseSearchResult } from './SearchResult';
 
 // This is how long we should wait after each keypress before actually executing the search
