+++ components/search/SearchButton.tsx
@@ -1,9 +1,7 @@
 import { useModal } from 'async-modals';
-import classNames from 'classnames';
-import { useRouter } from 'next/router';
-import React, { HTMLAttributes, useEffect } from 'react';
+import React, { HTMLAttributes } from 'react';
 import { FiSearch } from 'react-icons/fi';
-import IconButton from './atom/IconButton';
+import IconButton from '../atom/IconButton';
 import SearchModal from './SearchModal';
 
 const SearchButton: React.FC<HTMLAttributes<HTMLElement>> = ({ className }) => {
