+++ pages/account/signin.tsx
@@ -8,7 +8,7 @@ import Button from '../../components/atom/Button';
 import Card from '../../components/atom/Card';
 import FormGroup from '../../components/atom/FormGroup';
 import Input from '../../components/atom/Input';
-import { AuthContext } from '../../components/CognitoAuthProvider';
+import { AuthContext } from '../../components/general/CognitoAuthProvider';
 
 interface FormFields {
   username: string;
